requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bytesbufio==1.0.3
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
numpy==1.23.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
torch==1.11.0
typing_extensions==4.2.0
urllib3==1.26.9
django-cors-headers==3.13.0